Started huge refactoring
