+++ src/modules/interface.go
@@ -1,0 +1,5 @@
+package modules
+
+type Formatter interface {
+	Format(string) string
+}
